Push AST away from operators and towards builtin functions.
In the long run, we'd like to have a mix of user-defined and built-in functions,
all handled consistently. For example, one should be able to map the `edges`
operator across a stream of nodes just as one can map a lambda expression.

This commit removes the `Nullary` and `Unary` types, and introduces the `Var`
type, representing the category that includes both built-in operators and
user-defined variables.