Implement 'map'.
This is not as nice as one would like, because so many things are syntactic
operators, not functions. For example, you can't write `edges map` to map
`edges` across a stream, because `edges` is a postfix operator, not a function.
You need to say `((|x| x edges) map)`, which is grotesque.